# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Final Project Hacktiv8",
        "title": "Hacktiv8 Final Project - Muhammad Reyhan",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Muhammad Reyhan",
            "email": "muh.reyhan@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/todo": {
            "get": {
                "description": "Get specific todo by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get specific todo by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "todo serarch by id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ToDo"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "put": {
                "description": "Update Todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update Todo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ToDo"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "Create new Todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Todo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ToDo"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete Todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Todo",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ToDo"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/todos": {
            "get": {
                "description": "Get All Todos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show all available todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ToDo"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/user": {
            "put": {
                "description": "Update User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "Add User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete User by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "model.ToDo": {
            "type": "object",
            "properties": {
                "Desc": {
                    "type": "string"
                },
                "DueDate": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "PersonInCharge": {
                    "type": "integer"
                },
                "Status": {
                    "type": "integer"
                },
                "Title": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "Name": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        }
    }
}